refactor: moved code about "NewProfileScreen" from component folder to screen folder
